Towards creating multi-scale testing images
