+++ napari/layers/image/experimental/__init__.py
@@ -2,3 +2,4 @@
 """
 from .octree_intersection import OctreeIntersection
 from .octree_level import OctreeLevel
+from .octree_tile_builder import create_tiles
