+++ napari/layers/image/experimental/_octree_multiscale_slice.py
@@ -1,0 +1,11 @@
+"""OctreeMultiscaleSlice class.
+
+For viewing one slice of a multiscale image using an octree.
+"""
+
+
+class OctreeMultiscaleSlice:
+    """View a slice of an multiscale image using an octree."""
+
+    def __init__(self):
+        pass
